% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.import.R
\name{data.import}
\alias{data.import}
\alias{importUKNAMetals}
\alias{importUKNAPAHs}
\title{Data import functions}
\usage{
importUKNAMetals(site, pollutant, year, ...)

importUKNAPAHs(site, pollutant, year, ...)
}
\arguments{
\item{site}{(character) The identifier code(s) for the site(s) the data
is to be downloaded from.}

\item{pollutant}{(numeric) The pollutant identifier(s) of the pollutant(s)
to download data for.}

\item{year}{(numeric) The year(s) to download data for.}

\item{...}{(other arguments) currently ignored.}
}
\value{
The import functions typically return the request data as a
\code{data.frame} if data is available for requested site/pollutant/year
combination. In some cases the \code{data.frame} may be set up
for use with use with other packages.

\code{importUKNAMetals}: import function for metal data from the UK
Non-automatic Heavy Metals Network.
Previously called the Urban/Industrial Heavy Metals Network and Rural
Heavy Metals Network.

See https://uk-air.defra.gov.uk/networks/network-info?view=metals for
details and source information.

\code{importUKNAPAHs}: import function for metal data from the UK
Non-automatic Polycyclic Aromatic Hydrocarbons (PAH) Network.

see https://uk-air.defra.gov.uk/networks/network-info?view=pah for details
and source information.
}
\description{
Some functions to import data from held in on-line data
archives.
}
\note{
These functions were developed while working on projects that
required multiple downloads of data from these sources. The code is
shared freely but with the caveat that I am not the data provider.
So:

(a) You should cite both this package and the data source (identified
for each function), and

(b) You should be aware that they only work for as long as the structure
of the data in the archive remains as it was when the code was written,
and neither I or your have any control in that...
}
