% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.code.R
\name{legacy.code}
\alias{legacy.code}
\alias{missing.code}
\alias{extractHTMLAttachmentFromMSG}
\alias{readAIMSHTMLAttachment}
\title{legacy code}
\description{
grey.area is a home for working code that does
not have a proper home yet.

If you were expecting to find code that is not here, it may
have been rehoused, replaced or removed.

Rehoused functions:

Now in Umoh Edemeka's aircraftOPS package:

extractHTMLAttachmentFromMSG.

readAIMSHTMLAttachment
}
