% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.tools.R
\name{misc.tools}
\alias{misc.tools}
\alias{combinations}
\alias{uncombine}
\title{grey.area misc.tools}
\usage{
combinations(..., link = "<<>>", data = NULL)

uncombine(x, n = 1, splt = "<<>>")
}
\arguments{
\item{...}{Unspecified or additional arguments. For
\code{combinations} and \code{uncombine} vectors to
be combined or uncombined, respectively.}

\item{link}{(character) For \code{combinations}, the
string to use as a separator when combining vectors.}

\item{data}{(data.frame or similar) If supplied,
the source for vectors to be combined.}

\item{x}{(vector) For \code{uncombine}, a vector of
previously combined terms to be uncombined.}

\item{n}{(numeric) For \code{uncombine}, the combined
vector to be extracted. The default \code{1} extracts
the first vector of combined terms.}

\item{splt}{(character) For \code{uncombine}, the
separator previously used to combine vectors, used to
split vectors when uncombining terms.}
}
\value{
\code{combinations} and \code{uncombine} return
vectors of combined or uncombined terms, respectively.
}
\description{
Some misc functions.
}
\note{
These are misc functions. They might need
tidying

\code{combinations} merges supplied vectors but assumes
the terms in the supplied vectors are equivalent. This
is a little of a specialist function but if you ever
have a case where e.g. you want to group journeys
between locations need a-to-b and b-to-a to be merged
as a single case journeys-between-a-and-b...

\code{uncombine} extracts terms from a previous combined
vector.
}
\examples{
a <- 1:5
b <- 5:1
#pasting combines two vectors
#making  e.g. 1 + 5 and 5 + 1 unique cases
paste(a,b)
#combinations makes cases where these are equivalent
combinations(a,b)
}
\author{
Katrina Hemingway and Karl Ropkins
}
\keyword{methods}
