% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa.sunrise.code.R
\name{noaa.sunrise.code}
\alias{noaa.sunrise.code}
\alias{calcNOAASunrise}
\alias{noaa.sunrise}
\alias{cutDaylight}
\alias{hour.offset}
\title{NOAA Sunrise Calculations}
\usage{
calcNOAASunrise(
  time.stamp,
  latitude = 0,
  longitude = 0,
  local.hour.offset = 0,
  output = "all"
)

cutDaylight(mydata, latitude = 0, longitude = 0, local.hour.offset = 0)

hour.offset(local, pmatch = TRUE, ignore.case = TRUE)
}
\arguments{
\item{time.stamp}{(POSIX) A vector of date/time of
\code{POSIX*} class.}

\item{latitude, longitude}{(Numerics) The latitude and
longitude of the location where the associated \code{time.stamp}
measurement was logged. (Note: \code{latitude} + to North; - to
South. \code{longitude} + to East; - to West.)}

\item{local.hour.offset}{(Numeric or character) The associated
local time offset relative to GMT in hours (numeric) or a reference
term that \code{\link{hour.offset}} can convert into a numeric time
offset (character). (Note: + to East; - to West).}

\item{output}{(Character) Either the default \code{"all"}, which
returns all inputs and calculated parameters as a data.frame, or a
character vector of field/column names specifying which fields/columns
to return. See Value below for summary of calculated parameters
generated by this function.}

\item{mydata}{(Data.frame). A date frame which includes the
\code{time.stamp} vector as a field/column called \code{date}.}

\item{local}{(Character) The location (or reference) for required
hour offset.}

\item{pmatch, ignore.case}{(Logicals) Supplied \code{local} handling.
\code{pmatch} applies partial (leading edge) matching of reference
\code{local} terms. \code{ignore.case} resets \code{local} to upper
case before comparing with reference \code{local} terms. Both defaults
\code{TRUE}.}
}
\value{
\code{calcNOAASunrise} returns a data frame which, depending
on the call \code{output} value, may include the call arguments and
any of the following calculated parameters:
\describe{
\item{julian.day }{(Numeric) The Julian day.}
\item{julian.century }{(Numeric) The Julian century.}
\item{geom.mean.long.sun.deg }{(Numeric) The geometric mean longitude
of the Sun in degrees.}
\item{geom.mean.anom.sun.deg }{(Numeric) The geometric mean anomaly of
the Sun in degrees.}
\item{eccent.earth.orbit }{(Numeric) The eccentricity of Earth`s orbit.}
\item{sun.eq.of.ctr }{(Numeric) The estimated center for the Sun, in
degrees.}
\item{sun.true.long.deg }{(Numeric) The true longitude of the Sun, in
degrees.}
\item{sun.true.anom.deg }{(Numeric) The true anamoly of the Sun, in
degrees.}
\item{sun.rad.vector.aus }{(Numeric) The distance to the Sun, in AUs
(radial).}
\item{sun.app.long.deg }{(Numeric) The apparent longitude of the Sun,
in degrees.}
\item{mean.obliq.ecliptic.deg }{(Numeric) The mean obliquity of the
ecliptic, in degrees.}
\item{obliq.corr.deg }{(Numeric) The corrected obliquity of the
ecliptic, in degrees.}
\item{sun.rt.ascen.deg }{(Numeric) The right ascension of the Sun,
in degress.}
\item{sun.declin.deg }{(Numeric) The declination of the Sun, in
degrees.}
\item{vary }{(Numeric) TO BE CONFIRMED}
\item{eq.of.time.minutes }{(Numeric) TO BE CONFIRMED}
\item{ha.sunrise.deg }{(Numeric) TO BE CONFIRMED}
\item{solar.noon.lst }{(Numeric) TO BE CONFIRMED}
\item{sunrise.time.lst }{(Numeric) The estimated sunrise time as
fraction of day.}
\item{sunset.time.lst }{(Numeric) The estimated sunset time as
fraction of day.}
\item{sunlight.duration.minutes }{(Numeric) TO BE CONFIRMED}
\item{true.solar.time.min }{(Numeric) TO BE CONFIRMED}
\item{hour.angle.deg }{(Numeric) TO BE CONFIRMED}
\item{solar.zenith.angle.deg }{(Numeric) TO BE CONFIRMED}
\item{solar.elevation.angle.deg }{(Numeric) TO BE CONFIRMED}
\item{approx.atmospheric.refraction.deg }{(Numeric) TO BE CONFIRMED}
\item{solar.elevation.corrected.for.atm.refraction.deg }{(Numeric) TO
BE CONFIRMED}
\item{solar.azimuth.angle.deg.cw.from.n }{(Numeric) TO BE CONFIRMED}
\item{daylight }{(Factor) A factor specifying if the \code{time.stamp},
\code{latitude}, \code{longitude} and \code{local.hour.offset} is
estimated to be a \code{daylight} or \code{nighttime} instance.}
}
\code{cutDaylight} is an alternative to \code{calcNOAASunrise} which
accepts the time stamp as part of a data frame and returns that data
frame with the \code{calcNOAACalc} output component \code{daylight}
as an addition field/column.

\code{hour.offset} returns the hour.offset as a numeric.
}
\description{
Methods for the calculation of sunrise
and sunset times based on NOAA methods.
}
\details{
\code{calcNOAASunrise} is based on NOAA methods and calculates
a range of parameters based on the relative position of the Sun and a
receptor point of the Earth`s surface (see value below). The method is
valid for dates between 1901 and 2099.\code{cutDaylight} is a cut-down
version of the same method that was used in the \code{openair} package.

\code{hour.offset} uses \code{\link{ref.hour.offset}} as a look-up
table to match supplied \code{local} terms and return
\code{hour.offset} values for use in \code{\link{calcNOAASunrise}}
or \code{\link{cutDaylight}}.
}
\note{
The NOAA Sunrise methods were originally recoded for R as part
of work on the NERC penair project:
\url{http://www.openair-project.org/}.
\code{cutDaylight} is an alternative version of the function. It is
highly similar to code used as part of the function \code{cutData} in
the package \code{openair} to provide \code{daylight} conditioning for
many plot types within that package.

Character handling for \code{local.hour.offset} is by
\code{hour.offset} and \code{\link{ref.hour.offset}}.
}
\examples{
#example 1
#calcNOAASunrise
d <- seq(ISOdatetime(2000,1,31,0,0,0), by = "hour", length.out = 24)
ans <- calcNOAASunrise(d)
#solar elevation vs hour of day at (0,0) on 2000-01-31
plot(solar.elevation.corrected.for.atm.refraction.deg~time.stamp,
     main="2000-01-31; latitude 0, longitude 0",
     data=ans, type="b")

#anotate further, add dusk and dawn
#NB: sunrise and sunset are given as fractions of day
dusk <- ISOdatetime(2000,1,31,0,0,0) + (ans$sunset.time.lst*86400)
dawn <- ISOdatetime(2000,1,31,0,0,0) + (ans$sunrise.time.lst*86400)

abline(v=dusk, lty=2)
abline(v=dawn, lty=2)
abline(h=0, lty=3)
arrows(dusk, 25, dusk, 0, col="blue")
text(dusk, 25, "dusk", col="blue")
arrows(dawn, 25, dawn, 0, col="blue")
text(dawn, 25, "dawn", col="blue")
}
\references{
Methods based on:
\url{http://www.srrb.noaa.gov/highlights/sunrise/sunrise.html}
}
\seealso{
\code{\link{ref.hour.offset}}
}
\author{
Karl Ropkins
}
