% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.tools.R
\name{packaging.tools}
\alias{packaging.tools}
\alias{splat_pack}
\alias{splate_description}
\alias{splat_citation}
\alias{splat_package_intro}
\alias{splat_package_readme}
\alias{splat_package_new}
\alias{splat_function}
\alias{splat_data}
\alias{splat_date}
\alias{splat_version}
\alias{splat_description}
\alias{splat_package_news}
\title{packaging.tools}
\usage{
splat_pack(pkg = ".", overwrite = FALSE, pkg.dir = ".", ...)

splat_description(
  title = ".",
  description = ".",
  pkg = ".",
  overwrite = FALSE,
  ...
)

splat_citation(overwrite = FALSE, pkg = ".", ...)

splat_package_intro(
  name = ".",
  pkg = ".",
  overwrite = FALSE,
  title = ".",
  description = ".",
  ...
)

splat_package_readme(
  name = ".",
  pkg = ".",
  overwrite = FALSE,
  title = ".",
  description = ".",
  ...
)

splat_package_news(
  name = ".",
  pkg = ".",
  overwrite = FALSE,
  title = ".",
  description = ".",
  ...
)

splat_function(
  fun = ".",
  overwrite = FALSE,
  pkg = ".",
  title = ".",
  name = ".",
  ...
)

splat_data(data = ".", overwrite = FALSE, pkg = ".", ...)

splat_date(title = ".", date = ".", pkg = ".", overwrite = FALSE, ...)

splat_version(
  title = ".",
  version = ".",
  increase = ".",
  pkg = ".",
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{pkg}{(character) package name.}

\item{overwrite}{(logical) Overwrite file if already present,
default typically FALSE.}

\item{pkg.dir}{(character) main packaging directory.}

\item{...}{Other arguments passed on.}

\item{title}{(character) title of package or document template
built as part of splat.}

\item{description}{(character) text for package description.}

\item{name}{(character) name of document template if different from
item (e.g. function) being added to package.}

\item{fun}{(character) name of function to add to package.}

\item{data}{(character) name of data to add to package.}

\item{date}{(character) date in 'YYYY-MM_DD' format.}

\item{version}{(character) version in 'n.n.n' (or 'n.n.n.n') format.}

\item{increase}{(character or numeric) version increase, typically
character in 'n.n.n' (or 'n.n.n.n') format, but also allows the
numerics 1, 2, 3 or 4 as shorthand for increases of '1', '0.1',
'0.0.1' or '0.0.0.1', respectively.}
}
\description{
Code written when making packages.
}
