% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakInt.R
\name{peakInt}
\alias{peakInt}
\alias{plot.PeakInt}
\alias{print.PeakInt}
\title{peakInt}
\usage{
peakInt(
  x,
  peak.wd.asym = 1000,
  peak.ht.asym = 1000,
  peak.cluster = 0,
  peak.cutoff = "none",
  peak.cutoff.value = 0,
  ...
)

\method{plot}{PeakInt}(
  x,
  report.lim = c(1, -1),
  peak.col = "grey",
  peak.border = NA,
  data.col = "black",
  ...
)

\method{print}{PeakInt}(x, ...)
}
\arguments{
\item{peak.wd.asym}{peak property, peak width asymmetry}

\item{peak.ht.asym}{peak property, peak ht asymmetry}

\item{peak.cluster}{peak property, peak clustering}

\item{peak.cutoff}{peak property, peak cutoff}

\item{peak.cutoff.value}{peak property, peak cutoff value}

\item{...}{other arguments}

\item{(for}{peakInt) x numeric vector, typically a time-series of
data; (for PeakInt methods) an object of class 'PeakInt'}
}
\description{
peak integration
}
