% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.hour.offset.R
\docType{data}
\name{ref.hour.offset}
\alias{ref.hour.offset}
\title{Local Time Offsets}
\format{
A `data.frame` 38 locations and associated time offsets:
\describe{
\item{local}{(Character) The location for which an offset value is
reported.}
\item{hour.offset}{(Numeric) The time offset in hours associated
with \code{local}.}
}
}
\source{
NOAA assignment table:
http://www.srrb.noaa.gov/highlights/sunrise/timezone.html.
}
\usage{
ref.hour.offset
}
\description{
`ref.hour.offset` is a lookup table of local time
offsets.
}
\details{
\code{ref.hour.offset} is a lookup table for use with
\code{\link{calcNOAASunrise}} and \code{\link{hour.offset}}.
}
\examples{
#basic structure
head(ref.hour.offset)
#usage
hour.offset("greenwich")
hour.offset("india")
hour.offset("hawaii")
}
\keyword{datasets}
