% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanwal.code.R
\name{sanwal.code}
\alias{sanwal.code}
\alias{makeAQArchives}
\alias{makeArchive}
\alias{getArchiveDir}
\alias{getFromArchive}
\alias{testInArchive}
\alias{archiveThis}
\alias{mergeAQAndMet}
\title{sanwal.code}
\usage{
makeAQArchives(pass = "")

makeArchive(type = NULL)

getArchiveDir(type, pass = "")

getFromArchive(code, type, pass = "code")

testInArchive(code, type, pass)

archiveThis(obj, code, type)

mergeAQAndMet(aq, met)
}
\arguments{
\item{pass}{(character) function output control; what to do
if something unexpected happens.}

\item{type}{(character) data type, aq, met, merge or misc.}

\item{code}{(character) the names of the dataset to select,
typically the code of the aq data.}

\item{obj}{(data.frame or similar) the dataset, if, e.g., to
archive if archiving data.}

\item{aq}{(data.frame or similar) a dataset of air quality data.}

\item{met}{(data.frame or similar) a dataset of meteorological
data.}
}
\description{
Code written for Sanwal for archive build for
data analysis.
}
