% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracePlotOnImage.R
\name{tracePlotOnImage}
\alias{tracePlotOnImage}
\title{tracePlotOnImage}
\usage{
tracePlotOnImage(n = 1000)
}
\arguments{
\item{n}{(integer) the maximum number of data points to collect
when tracing information on a plotted image. The option
\code{n = -1} allows open ended data collection.}
}
\description{
Written for use in Muchammad Nafiek MSc project,
\code{tracePlotOnImage} allows you to trace information in a
image file.

\code{tracePlotOnImage} allows the user to trace
information on an image, e.g. a picture file containing a bar graph
or plot, using the mouse left button. The procedure has three
steps: (1) The user is asked to select the the bottom corner of the
plot and enter its coordinates; (2) The user is asked to select the
top corner of the plot and enter its coordinates; and (3) The user
is asked to identify points of interest on the plot. The user can
select \code{n} points or stop the point collection process early
by pressing the mouse right button and selecting exit. The data
points, calibrated using information from steps (1) and (2), are
then returned.
}
\author{
Karl Ropkins (with contributions from Muchammad Nafiek)
}
