% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grey.area-package.R
\docType{package}
\name{grey.area}
\alias{grey.area}
\title{grey.area: This is a bit of a grey area...}
\description{
This is an in-house R package.
}
\details{
The original package was intended as somewhere to store
code that written on projects but not properly developed
for third-party use outside those projects.

This happens a lot with projects. The code has to be good
enough to do the project in hand, but there is no time or
need to tidy the code, make it user-friendly, or document
it properly.

This package update is an attempt to tidy the sprawl of the
previous version of the package...

Please contact Karl Ropkins for details.
}
\section{grey.area functions}{

Packaged Functions include: geo.tools, packaging.tools
}

